The arguments of the FxT traces are of types "unsigned long". Fix a typo in the trace generation too.

+++ include/starpu_util.h
@@ -141,7 +141,7 @@ static inline int starpu_get_env_number(const char *str)
 }
 
 /* Add an event in the execution trace if FxT is enabled */
-void starpu_trace_user_event(unsigned code);
+void starpu_trace_user_event(unsigned long code);
 
 /* Some helper functions for application using CUBLAS kernels */
 void starpu_helper_cublas_init(void);
